forras

+++ src/main/java/hu/tmx/mybooksapp/entity/User.java
@@ -2,11 +2,10 @@ package hu.tmx.mybooksapp.entity;
 
 import lombok.*;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 import javax.validation.constraints.*;
+import java.util.HashSet;
+import java.util.Set;
 
 @Entity(name="users")
 @NoArgsConstructor
@@ -18,7 +17,7 @@ public class User {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     @Setter(AccessLevel.NONE)
-    private Long id;
+    private int id;
     @NotNull
     @Email
     @Pattern(regexp=".+@.+\\..+")
